Pyannote+Whisper/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
autopage==0.5.1
backports.cached-property==1.0.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
cytoolz==0.12.0
dataclasses==0.6
decorator==5.1.1
docopt==0.6.2
einops==0.3.2
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
hmmlearn==0.2.8
huggingface-hub==0.11.1
HyperPyYAML==1.1.0
idna==3.4
importlib-metadata==4.13.0
intervaltree==3.1.0
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lhotse==1.10.0
librosa==0.9.2
lilcom==1.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.8
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
optuna==3.0.4
packaging==21.3
pandas==1.5.2
pbr==5.11.0
Pillow==9.3.0
pooch==1.6.0
prettytable==3.5.0
primePy==1.3
protobuf==3.20.1
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytorch-metric-learning==1.6.3
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.8.1
semver==2.13.0
sentencepiece==0.1.97
shellingham==1.5.0
simplejson==3.18.0
singledispatchmethod==1.0
six==1.16.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
speechbrain==0.5.13
SQLAlchemy==1.4.44
stevedore==4.1.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.0
transformers==4.24.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@eff383b27b783e280c089475852ba83f20f64998
yarl==1.8.1
zipp==3.11.0
